@@ -1,6 +1,6 @@
 {
   "name": "angular-boilerplate",
-  "version": "0.0.0",
+  "version": "5.0.0",
   "private": true,
   "scripts": {
     "prepare": "npx simple-git-hooks",
@@ -33,43 +33,43 @@
     ]
   },
   "dependencies": {
-    "@angular/animations": "^16.0.0",
-    "@angular/common": "^16.0.0",
-    "@angular/compiler": "^16.0.0",
-    "@angular/core": "^16.0.0",
-    "@angular/forms": "^16.0.0",
-    "@angular/platform-browser": "^16.0.0",
-    "@angular/platform-browser-dynamic": "^16.0.0",
-    "@angular/router": "^16.0.0",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
+    "@angular/animations": "^16.0.2",
+    "@angular/cdk": "^16.0.1",
+    "@angular/common": "^16.0.2",
+    "@angular/compiler": "^16.0.2",
+    "@angular/core": "^16.0.2",
+    "@angular/forms": "^16.0.2",
+    "@angular/platform-browser": "^16.0.2",
+    "@angular/platform-browser-dynamic": "^16.0.2",
+    "@angular/router": "^16.0.2",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.5.2",
     "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.0.1",
+    "@angular-devkit/build-angular": "^16.0.2",
     "@angular-eslint/builder": "16.0.1",
     "@angular-eslint/eslint-plugin": "16.0.1",
     "@angular-eslint/eslint-plugin-template": "16.0.1",
     "@angular-eslint/schematics": "16.0.1",
     "@angular-eslint/template-parser": "16.0.1",
-    "@angular/cli": "~16.0.1",
-    "@angular/compiler-cli": "^16.0.0",
-    "@culur/stylelint-config": "^1.2.0",
-    "@iconify/json": "^2.2.64",
+    "@angular/cli": "~16.0.2",
+    "@angular/compiler-cli": "^16.0.2",
+    "@iconify/json": "^2.2.67",
     "@iconify/tailwind": "^0.1.2",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/typography": "^0.5.9",
-    "@types/jasmine": "~4.3.0",
-    "@types/node": "^20.1.3",
+    "@types/jasmine": "~4.3.1",
+    "@types/node": "^20.2.1",
     "@typescript-eslint/eslint-plugin": "5.59.2",
     "@typescript-eslint/parser": "5.59.2",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.39.0",
+    "eslint": "^8.41.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
     "jasmine-core": "~4.6.0",
-    "karma": "~6.4.0",
+    "karma": "~6.4.2",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
@@ -80,8 +80,9 @@
     "prettier-eslint": "^15.0.1",
     "prettier-plugin-tailwindcss": "^0.2.8",
     "simple-git-hooks": "^2.8.1",
-    "stylelint": "^15.6.1",
+    "stylelint": "^15.6.2",
+    "stylelint-config-standard": "^33.0.0",
     "tailwindcss": "^3.3.2",
-    "typescript": "~5.0.2"
+    "typescript": "~5.0.4"
   }
 }
