@@ -1,90 +1,112 @@
 import { DOCUMENT } from '@angular/common';
-import { Injectable, OnDestroy, inject } from '@angular/core';
-import { DEFAULT_BASE_THEME } from '@lib/constants';
+import { Injectable, computed, effect, inject, signal } from '@angular/core';
+import { toSignal } from '@angular/core/rxjs-interop';
+import { STORAGE } from '@lib/constants';
 import { storage } from '@lib/utils';
-import { BehaviorSubject, Subject, fromEventPattern } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
-import { AppTheme } from './theme.config';
+import { fromEventPattern } from 'rxjs';
+import { ThemeFontSize, ThemePrimaryColor, ThemeSchema } from './theme.config';
 
 @Injectable({
     providedIn: 'root',
 })
-export class ThemeService implements OnDestroy {
-    currentTheme$ = new BehaviorSubject<AppTheme | null>(this._storedTheme);
-
+export class ThemeService {
     private readonly _document = inject(DOCUMENT);
-
-    private readonly _destroy$ = new Subject();
 
     private readonly _mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
 
-    public get currentTheme(): AppTheme | null {
-        return this.currentTheme$.getValue();
-    }
-
-    public get systemTheme(): AppTheme {
-        return this._mediaQuery.matches ? 'dark' : 'light';
-    }
-
-    private get _storedTheme(): AppTheme | null {
-        return storage.getItem('appTheme');
-    }
-
-    private set _storedTheme(theme: AppTheme | null) {
-        storage.setItem('appTheme', theme as AppTheme);
-    }
-
-    ngOnDestroy(): void {
-        this._destroy$.complete();
-        this._destroy$.unsubscribe();
-    }
-
-    init(): void {
-        this.setTheme(this._storedTheme || DEFAULT_BASE_THEME);
-        this._listenForMediaQueryChanges();
-    }
-
-    /**
-     * Manually changes theme in LocalStorage & HTML body
-     *
-     * @param theme new theme
-     */
-    setTheme(theme: AppTheme): void {
-        this._clearThemes();
-        this._storedTheme = theme;
-
-        let bodyClass = theme;
-        this.currentTheme$.next(bodyClass);
-
-        if (theme === 'system') {
-            bodyClass = this.systemTheme;
-        }
-        this._document.body.classList.add(bodyClass);
-    }
-
-    /**
-     * Handles system theme changes & applies theme automatically
-     *
-     */
-    private _listenForMediaQueryChanges(): void {
+    private readonly _mediaChanges = toSignal(
         fromEventPattern<MediaQueryListEvent>(
             this._mediaQuery.addListener.bind(this._mediaQuery),
             this._mediaQuery.removeListener.bind(this._mediaQuery),
-        )
-            .pipe(takeUntil(this._destroy$))
-            .subscribe(() => {
-                // Only applies changes when the current theme is "system"
-                if (this._storedTheme === 'system') {
-                    this.setTheme('system');
-                }
-            });
+        ),
+    );
+
+    private readonly _systemTheme = computed(() => (this._mediaQuery.matches ? ThemeSchema.Dark : ThemeSchema.Light));
+
+    schema = signal(this._defaultSchema);
+    fontSize = signal(this._defaultFontSize);
+    primaryColor = signal(this._defaultPrimaryColor);
+
+    get availableSchemas(): ThemeSchema[] {
+        return Object.values(ThemeSchema);
     }
 
-    /**
-     * Clears all themes in ThemeList enum from the HTML element
-     *
-     */
-    private _clearThemes(): void {
-        this._document.body.classList.remove('system', 'light', 'dark');
+    get availableFontSizes(): ThemeFontSize[] {
+        return Object.values(ThemeFontSize);
+    }
+
+    get availablePrimaryColors(): ThemePrimaryColor[] {
+        return Object.values(ThemePrimaryColor);
+    }
+
+    private get _defaultSchema(): ThemeSchema {
+        return storage.getItem<ThemeSchema>(STORAGE.theme.schema) ?? ThemeSchema.System;
+    }
+
+    private get _defaultFontSize(): ThemeFontSize {
+        return storage.getItem<ThemeFontSize>(STORAGE.theme.fontSize) ?? ThemeFontSize.Md;
+    }
+
+    private get _defaultPrimaryColor(): ThemePrimaryColor {
+        return storage.getItem<ThemePrimaryColor>(STORAGE.theme.primaryColor) ?? ThemePrimaryColor.Lime;
+    }
+
+    constructor() {
+        effect(() => {
+            if (this._mediaChanges() && this.schema() === ThemeSchema.System) {
+                this.setSchema(ThemeSchema.System);
+                storage.setItem(STORAGE.theme.schema, this.schema());
+            }
+
+            if (this.schema()) {
+                storage.setItem(STORAGE.theme.schema, this.schema());
+            }
+
+            if (this.fontSize()) {
+                storage.setItem(STORAGE.theme.fontSize, this.fontSize());
+            }
+
+            if (this.primaryColor()) {
+                storage.setItem(STORAGE.theme.primaryColor, this.primaryColor());
+            }
+        });
+    }
+
+    initialize(): void {
+        this.setSchema(this._defaultSchema);
+        this.setFontSize(this._defaultFontSize);
+        this.setPrimaryColor(this._defaultPrimaryColor);
+    }
+
+    reset(): void {
+        this.setSchema(ThemeSchema.System);
+        this.setFontSize(ThemeFontSize.Md);
+        this.setPrimaryColor(ThemePrimaryColor.Lime);
+    }
+
+    setSchema(schema: ThemeSchema): void {
+        for (const theme of Object.values(ThemeSchema)) {
+            this._document.documentElement.classList.remove(theme);
+        }
+
+        this.schema.set(schema);
+
+        if (schema === ThemeSchema.System) {
+            this._document.documentElement.classList.add(this._systemTheme());
+        } else {
+            this._document.documentElement.classList.add(schema);
+        }
+    }
+
+    setFontSize(fontSize: ThemeFontSize): void {
+        this.fontSize.set(fontSize);
+
+        this._document.documentElement.setAttribute('data-font-size', fontSize);
+    }
+
+    setPrimaryColor(primaryColor: ThemePrimaryColor): void {
+        this.primaryColor.set(primaryColor);
+
+        this._document.documentElement.setAttribute('data-primary-color', primaryColor);
     }
 }
