@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { FooterComponent, NavbarComponent } from '@lib/components';
+import { FooterComponent } from '@lib/components/footer/footer.component';
+import { NavbarComponent } from '@lib/components/navbar/navbar.component';
 
 @Component({
     selector: 'app-layout-horizontal',
