@@ -1,24 +1,40 @@
-import { Injectable } from '@angular/core';
+import { Injectable, computed, effect, signal } from '@angular/core';
+import { STORAGE } from '@lib/constants';
+import { User } from '@lib/interfaces';
 import { storage } from '@lib/utils/storage/storage.utils';
-import { BehaviorSubject } from 'rxjs';
 
 @Injectable({
     providedIn: 'root',
 })
 export class AuthService {
-    isAuthenticated$ = new BehaviorSubject<boolean>(!!storage.getItem('appSession'));
+    session = signal<{ token: string; user: User } | null>(storage.getItem(STORAGE.session));
 
-    get isAuthenticated(): boolean {
-        return this.isAuthenticated$.getValue();
+    isAuthenticated = computed(() => !!this.session());
+
+    constructor() {
+        effect(() => {
+            const { token, user } = this.session() || {};
+
+            if (token && user) {
+                storage.setItem(STORAGE.session, { token, user });
+            } else {
+                storage.removeItem(STORAGE.session);
+            }
+        });
     }
 
     login(): void {
-        storage.setItem('appSession', { user: 'some-user-id', token: 'abc' });
-        this.isAuthenticated$.next(true);
+        this.session.set({
+            token: 'abc',
+            user: {
+                username: 'johnwick2097',
+                email: 'johnwick@mail.com',
+                role: 'admin',
+            },
+        });
     }
 
     logout(): void {
-        storage.removeItem('appSession');
-        this.isAuthenticated$.next(false);
+        this.session.set(null);
     }
 }
