@@ -1,1 +1,26 @@
-export type AppTheme = 'system' | 'light' | 'dark';
+export enum ThemeSchema {
+    System = 'system',
+    Light = 'light',
+    Dark = 'dark',
+}
+
+export enum ThemeFontSize {
+    Sm = 'sm',
+    Md = 'md',
+    Lg = 'lg',
+}
+
+export enum ThemePrimaryColor {
+    Green = 'green',
+    Teal = 'teal',
+    Lime = 'lime',
+    Blue = 'blue',
+    Indigo = 'indigo',
+    Purple = 'purple',
+    Violet = 'violet',
+    Yellow = 'yellow',
+    Amber = 'amber',
+    Pink = 'pink',
+    Red = 'red',
+    Rose = 'rose',
+}
