@@ -1,5 +1,3 @@
-import { StorageObjectData, StorageObjectType } from './storage.types';
-
 type StorageOptions = {
     api?: 'LocalStorage' | 'SessionStorage';
 };
@@ -8,28 +6,24 @@ function getStorageApi(api: StorageOptions['api']): Storage {
     return api === 'SessionStorage' ? sessionStorage : localStorage;
 }
 
-function getItem<T extends StorageObjectType>(item: T, options?: StorageOptions): StorageObjectData<T>['data'] | null {
+function getItem<T>(name: string, options?: StorageOptions): T | null {
     const api = getStorageApi(options?.api || 'LocalStorage');
-    const data = api.getItem(item.toString());
-    return data ? (JSON.parse(data) as StorageObjectData<T>['data']) : null;
+    const data = api.getItem(name);
+    return data ? (JSON.parse(data) as T) : null;
 }
 
-function setItem<T extends StorageObjectType>(
-    itemName: T,
-    data: StorageObjectData<T>['data'],
-    options?: StorageOptions,
-): void {
+function setItem<T>(name: string, data: T, options?: StorageOptions): void {
     if (data === null || data === undefined) {
         return;
     }
 
     const api = getStorageApi(options?.api || 'LocalStorage');
-    api.setItem(itemName, JSON.stringify(data));
+    api.setItem(name, JSON.stringify(data));
 }
 
-function removeItem<T extends StorageObjectType>(item: T, options?: StorageOptions): void {
+function removeItem(name: string, options?: StorageOptions): void {
     const api = getStorageApi(options?.api || 'LocalStorage');
-    api.removeItem(item);
+    api.removeItem(name);
 }
 
 function clear(options?: StorageOptions): void {
